requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate@c7b36255925267349fd25341a21a08a9856d06fd
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
Authlib==1.6.1
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
azure-ai-agents==1.1.0b1
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b11
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
blake3==1.0.5
bottle==0.13.4
cachetools==6.1.0
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
coverage==7.6.12
cryptography==45.0.3
cupy-cuda12x==13.5.1
cyclopts==3.22.5
datasets==3.3.2
decorator==5.2.1
deepspeed==0.17.4
Deprecated==1.2.18
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastmcp==2.11.1
fastrlock==0.8.3
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
interegular==0.3.3
isodate==0.7.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
langchain-core==0.3.63
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.43
lark==1.2.2
latex2sympy2_extended==1.10.2
lazy-object-proxy==1.11.0
lazy_loader==0.4
librosa==0.10.2.post1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.8.0
mcp==1.12.3
mdurl==0.1.2
mistral_common==1.8.3
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.4.8
openai==1.90.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python-headless==4.12.0.88
opentelemetry-api==1.33.1
orjson==3.10.18
ormsgpack==1.10.0
outlines_core==0.2.10
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
partial-json-parser==0.2.1.1.post6
pathable==0.4.4
peft @ git+https://github.com/huggingface/peft@f51203f3e4d8b3856596d3296dae7180f15910c1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyperclip==1.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.1
ray==2.48.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-rst==1.3.1
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==3.0.2
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
trl @ git+https://github.com/huggingface/trl@491921c1a4167e7c84429382470b0bb3158e66b0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.18.3
zipp==3.22.0
zstandard==0.23.0
